Move RefferralModal to root components folder
