+++ src/components/RefferralModal/index.ts
@@ -1,0 +1,2 @@
+export { default } from './RefferralModal'
+export * from './components'
