+++ src/pages/common/LeaderboardPage/components/ReferralTab/ReferralTab.module.less
@@ -120,164 +120,3 @@
   min-width: 128px;
 }
 // * ReferralInfoSection * //
-
-// * ReferralInput * //
-.referralInputField {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-.referralInputLabel {
-  color: var(--content-secondary);
-  font: var(--title-sm);
-  text-transform: uppercase;
-
-  @media (max-width: 640px) {
-    font: var(--title-xs);
-  }
-}
-
-.referralInputWrapper {
-  position: relative;
-}
-
-.referralInput {
-  border: var(--border-primary) !important;
-}
-
-.referralInputButton {
-  background: var(--bg-primary);
-  border: var(--border-primary);
-  border-radius: 4px;
-  cursor: pointer;
-
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  height: 24px;
-
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  right: 4px;
-
-  span {
-    font: var(--btn-text-sm);
-  }
-}
-// * ReferralInput * //
-
-// * RefferralModal * //
-.modal :global {
-  .ant-modal-body {
-    padding: 0 !important;
-  }
-
-  .ant-modal-content {
-    padding: 0;
-  }
-}
-
-.referralModalContent {
-  display: flex;
-  flex-direction: column;
-  padding: 24px 24px 0;
-  gap: 12px;
-
-  @media (max-width: 640px) {
-    padding: 16px 16px 0;
-    gap: 8px;
-  }
-
-  .title {
-    font: var(--title-xl);
-    text-align: center;
-    text-transform: uppercase;
-
-    @media (max-width: 640px) {
-      font: var(--title-lg);
-    }
-  }
-
-  .subtitle {
-    font: var(--body-text-md);
-    text-align: center;
-
-    @media (max-width: 640px) {
-      font: var(--body-text-sm);
-    }
-  }
-}
-
-.referralModalInfo {
-  display: flex;
-  align-items: center;
-}
-
-.banxImage {
-  margin-left: -24px;
-  height: 132px;
-  width: 132px;
-}
-
-.referralModalInfoRows {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.referralModalInfoRow {
-  background: var(--bg-primary);
-  border-radius: 12px;
-
-  display: flex;
-  align-items: center;
-  padding: 8px;
-  gap: 8px;
-
-  svg {
-    min-height: 24px;
-    min-width: 24px;
-  }
-
-  span {
-    font: var(--body-text-sm);
-  }
-}
-
-.referrerModalInfo {
-  padding: 12px 24px 24px;
-
-  @media (max-width: 640px) {
-    padding: 8px 16px 16px;
-  }
-}
-
-.referrerWallet {
-  font: var(--body-text-sm);
-
-  @media (max-width: 640px) {
-    font: var(--body-text-xs);
-  }
-}
-
-.confirmButton {
-  margin: 12px 0;
-  width: 100%;
-
-  @media (max-width: 640px) {
-    margin: 8px 0;
-  }
-}
-
-.warningMessage {
-  color: var(--content-secondary);
-  font: var(--body-text-sm);
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-// * RefferralModal * //
