+++ src/pages/common/RootPage/RootPage.tsx
@@ -1,17 +1,32 @@
+import { useEffect } from 'react'
+
 import classNames from 'classnames'
 import { NavLink } from 'react-router-dom'
 
 import { Button } from '@banx/components/Buttons'
+import RefferralModal from '@banx/components/RefferralModal'
 
-import { Theme, useTheme } from '@banx/hooks'
+import { Theme, extractReferralCodeFromPath, useTheme } from '@banx/hooks'
 import { InfinityIcon, Lightning, PencilLtv } from '@banx/icons'
 import { PATHS } from '@banx/router'
+import { useBanxLogin, useModal } from '@banx/store/common'
 
 import { Interest } from './icons'
 
 import styles from './RootPage.module.less'
 
 export const RootPage = () => {
+  const { open } = useModal()
+  const { jwt } = useBanxLogin()
+
+  useEffect(() => {
+    const referralCode = extractReferralCodeFromPath(location.pathname)
+
+    if (referralCode && !jwt) {
+      open(RefferralModal)
+    }
+  }, [jwt, open])
+
   return (
     <div className={styles.pageWrapper}>
       <Header />
